requirements.txt
datasets>=2.0.0
pandas>=1.5.0
huggingface_hub>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
networkx>=3.0 